Вносит запрошенные изменения
