+++ js/generate-comments.js
@@ -1,37 +1,5 @@
-import {getRandomInteger, shuffleArray, generateIntId, getArrayRandomElement} from './util.js';
-
-// Константы
-const COMMENT_SENTENCES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const COMMENT_AUTHOR_NAMES = [
-  'Марьяна',
-  'Олег',
-  'Никита',
-  'Елена',
-  'Борис',
-  'Ядвига',
-  'Кристина',
-  'Стефан',
-  'Роман',
-  'Григорий',
-  'Марк',
-  'Таисия',
-  'Егор',
-  'Андрей',
-  'Анеля'
-];
-
-const USER_AVATAR_MIN_ID = 1;
-const USER_AVATAR_MAX_ID = 6;
-const COMMENT_ID_START = 10;
-const COMMENT_ID_STEP = 5;
+import { getRandomInteger, shuffleArray, generateIntId, getArrayRandomElement } from './util.js';
+import {COMMENT_SENTENCES, COMMENT_AUTHOR_NAMES, USER_AVATAR_MIN_ID, USER_AVATAR_MAX_ID, COMMENT_ID_START, COMMENT_ID_STEP} from './constants.js';
 
 //Комменты
 const makeCommentMessage = function () {
